refactor: reorganize API services and
