+++ frontend/src/api/config/api-config.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { useAuthStore } from "../stores/auth.store";
+import { useAuthStore } from "../../stores/auth.store";
 
 // Environment detection
 const isDevelopment = import.meta.env.DEV;
